# crypto/des: Performance optimization of permutation function in DES algorithm