- Description:
Inspired by the function `func permuteInitialBlock(block uint64) uint64` (crypto/des/block.go, line 119) , I converted the function `c.subkeys[i] = unpack(permuteBlock(pc2Input, permutedChoice2[:]))` call on line 243 of crypto/des/block.go into a bit operation method. 

- Performance improvement (based on arm64-arch):
```bash
name          old time/op    new time/op    delta
DesCBCDemo-2    8.38µs ± 0%    5.97µs ± 0%  -28.69%

name          old alloc/op   new alloc/op   delta
DesCBCDemo-2    1.08kB ± 0%    1.08kB ± 0%    0.00%

name          old allocs/op  new allocs/op  delta
DesCBCDemo-2      18.0 ± 0%      18.0 ± 0%    0.00%
```
- Test demo:
```go
func RunDesKeyGenDemo() error {
	key := []byte("abcdefgh")
	_, err := des.NewCipher(key)
	if err != nil {
		return err
	}
	return nil
}

func DesCBCEncrypt(rawData, key []byte) ([]byte, error) {
	block, err := des.NewCipher(key)
	if err != nil {
		return nil, err
	}
	blockSize := block.BlockSize()
	rawData = crypto.PKCS5Padding(rawData, blockSize)
	cipherText := make([]byte, blockSize+len(rawData))
	iv := cipherText[:blockSize]
	if _, err := io.ReadFull(rand.Reader, iv); err != nil {
		panic(err)
	}
	blockMode := cipher.NewCBCEncrypter(block, iv)
	blockMode.CryptBlocks(cipherText[blockSize:], rawData)
	return cipherText, nil
}

func DesCBCDecrypt(encryptData, key []byte) ([]byte, error) {
	block, err := des.NewCipher(key)
	if err != nil {
		return nil, err
	}
	blockSize := block.BlockSize()
	if len(encryptData) < blockSize {
		return nil, errors.New("cipher Text length error")
	}
	iv := encryptData[:blockSize]
	encryptData = encryptData[blockSize:]
	if len(encryptData)%blockSize != 0 {
		return nil, errors.New("ciphertext is not a multiple of the block size")
	}
	blockMode := cipher.NewCBCDecrypter(block, iv)
	rawData := make([]byte, len(encryptData))
	blockMode.CryptBlocks(rawData, encryptData)
	rawData = crypto.PKCS5UnPadding(rawData)
	return rawData, nil
}
```